requirements.txt
altair==5.5.0
appnope==0.1.4
art==6.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.4.0
cairocffi==1.7.1
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.8
Django==5.0.7
django-extensions==3.2.3
django-taggit==5.0.1
executing==2.2.0
feature-engine==1.8.3
filelock==3.13.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.33.0
google-auth-oauthlib==1.2.1
graphqlclient==0.2.4
graphviz==0.20.3
gspread==6.1.2
idna==3.7
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
MouseInfo==0.1.3
narwhals==1.41.0
nest-asyncio==1.6.0
numpy==1.26.2
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.1.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.9.1
pydotplus==2.0.2
pyfiglet==1.0.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2023.3.post1
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9
rubicon-objc==0.5.0
scikit-learn==1.6.1
scipy==1.15.3
six==1.16.0
smmap==5.0.2
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2023.3
urllib3==2.2.2
virtualenv==20.25.0
wcwidth==0.2.13
